requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
bidict==0.21.2
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
idna==2.10
multidict==5.1.0
mypy-extensions==0.4.3
pathspec==0.8.1
pygame==2.0.1
python-engineio==4.1.0
python-socketio==5.2.1
regex==2021.3.17
requests==2.25.1
six==1.16.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
websocket-client==0.59.0
yarl==1.6.3
